requirements.txt
web3==6.11.1
aiohttp==3.9.1
python-dotenv==1.0.0
eth-account==0.10.0
eth-utils==2.3.0
eth-typing==3.5.1
requests==2.31.0
websockets==11.0.3
aiosignal==1.3.1
attrs==23.1.0
